+++ src/router/index.js
@@ -1,0 +1,13 @@
+import routerProductos from './routerProductos.js'
+import routerCarritos from './routerCarritos.js'
+import routerVistas from './routerVistas.js'
+import routerAuth from './routerAuth.js'
+import routerTest from './routerTest.js'
+
+export {
+    routerProductos,
+    routerCarritos,
+    routerVistas,
+    routerAuth,
+    routerTest
+}
